Refactor into phone_number.py and number_homoglyphs.py
